The problem I mentioned in my check-in earlier this week (for the fix for *.RAW file handling) is pretty bad, and I think we need a fix for it.
Problem is that the output file has the extension .raw which will clobber
their input data in win32.

I've done the simplest possible fix, which is to detect when the input and
output files are identical, and always add "_out" to the end of the output
basename in this situation.  It isn't ideal since the output file column
will still have the old name, but I think this is preferable, since the
user should really intervene and change the name manually.  But if they
don't (i.e., they ignore the "overwrite" warning), we'll cover for them.

I've surrounded the fix with "#ifdef win32", since this will really only
come up on windows.

I will put a new version of the windows build on seth, and re-test.