Another library to combine functionality. This one generates the interferogram and coherence in one wash and multilooks the interferogram. Avoids reading the complex data twice in a row.
Tested the results. Got identical binary results for single look
interferogram and coherence. The standalone multilook works with a sine/cosine
index table and uses a conversion back from amplitude and phase to
complex number in order to do the multilooking. Statistics for the
multilooked version looked very close though. 

Another issue was the mixed use of float versus double values for
internal calculation. Since somebody deemed it necessary to go to
double in the first place I adapted this to all calculations
used. This changes the results as well slightly.