Windows version can now actually *read* images, in addition to writing them. The problem here was, um, curious.
The exact symptoms are:
	- Everything works fine in Linux, IRIX, and Mac OS X.
	- Everything works fine from a Windows main program (not in a DLL)
	- fopen, fread, fwrite, and fclose work fine from a Windows DLL.
	- fseek in a DLL results in an exception 0xc0000005 (access violation)
	  because the memory at 0x00000010 could not be written by 
	  ntdll.dll's mutex lock routines.

The fix is to link all the involved DLLs with "/MD", which drags in
"msvcprt.lib", which drags in the multithreaded C runtime library.  
In the VC++ IDE, this option is marked "Use runtime library: Multithreaded DLL".

	http://support.microsoft.com/kb/140584/en-us
	http://msdn2.microsoft.com/en-us/library/2kzt1wy3.aspx

Obvious, right?  Right?!?  Only took 10 hours or so to find...

Why no sir, I don't think Windows DLLs are broken beyond repair
by design--certainly not!  I'm offended that you even suggest it!