Major reshuffling of prototype's meta code.
The *one* new thing is asf_meta/enum.h, and a perl script that parses
all the values for each "enum" listed in this header into a big
C-linkable table of string values.  I'm about to use this table to
parse the .meta file string fields, and it should come in handy for 
GUI and debugging display of metadata_source's values.