Incorporate refine_offset into terrain correction.
After the fftMatch, if the line/sample shifts are "too large" (currently
more than 2.5 pixels in either direction), then the metadata for the
slant range image is updated (in particular, the azimuthal time shift,
and the slant shift values) to eliminate this offset.  After this, the
DEM is re-cut and the fftMatch is performed a second time, after which
we should have an excellent match.  In the cases I've tried, the
matches are very good.

This is a re-write of the refine_offset tool, that uses the GSL 2-d
optimization to find the proper time and slant shifts.

Since the root finder is so finicky, I implemented a coarse root finder
to generate a starting point.  The coarse finder generates a grid over
the slant-shift/time-shift plane, finds the min, then repeats with a
smaller grid centered around this new min.

In many cases, it appears this coarse root finder finds a good enough
root, so that the GSL root finder doesn't refine the shifts at all.

I've left the hordes of printfs in there for now, while we decide if
what it is doing is sensible.

After the proper shifts are determined, the DEM is re-cut