Add support to libasf_terrcorr to not do any terrain correction - just applies the refine_offset code, in order to improve the geolocation of an image based on the dem.
Created a new program "refine_geolocation" to use this functionality.
(It seemed a little silly to have asf_terrcorr support an option like
"no-terrain-correct"!)  This program leaves the image alone, just
creates a new metadata file with the determined offsets to the azimuth
and slant shifts.  

We probably don't want to leave it like this, though - should it
modify the metadata file of the input image?  I guess this is better
than creating a whole new image (that is exactly the same) just for
a small metadata change... perhaps the best solution is to put the
original time/slant shifts in the metadata (as comments or something?)
so that the user can recover if the tool botches things...