Change getLook() to not abort when an error occurs, and many other functions to return an error code (non-zero on error, zero on success).  At the moment we just have a single error value: 1, but client code should not assume this will always be the case.
Most code ignores these return values (as the functions used to return
void), we should probably add in "assert(ret==0);" or something
similar in most places so that the behavior is similar to before.

This change was required by libasf_sar/refine_offset, which uses a 2-d
root finder to adjust slant and time offsets.  During the root
finding, it is likely that poor slant and time shifts will be tried,
including some that result in look vectors that don't hit the earth,
or other problems.  In these cases, we don't want the program to
abort, we just want to search for the proper values in the other
direction.  So, in those cases refine_offset now just sets the error
value to a big number.

I also tweaked the searching in refine_offset, so it should be a
little faster (it uses a coarser grid, and divides the search grid
more quickly).