A new feature: automatically masking out water.  What this does is allow the user to create a water mask automatically from the DEM -- all DEM values <=1 are tagged as masked.
The code in the tool "mask_from_dem" was used to produce the mask.

In "asf_terrcorr", this feature is activated via the command-line
option "-auto-water-mask".

In "asf_convert", this feature is controlled by the flag
"auto water mask" (with value either 1 or 0) in the configuration
file.

Some changes were made in the water masking code:
 - Dynamically compute size cutoffs
 - Added code to handle situations where seed squares can't be found
 - The seed square code returns 20 potential squares, the largest of
   these is selected and used in the fftMatch.  If the match has a
   certainty value smaller than a hard-coded cutoff (60% at the
   moment), the next-largest seed region is selected in an attempt
   to get a better match.  

   Wouldn't it be great to run the fftMatch on ALL of the 20 regions,
   and use the result with the "best" region size & fftMatch cert ??