Change to how the seed squares work.  Have a hard cutoff for the minimum size now, instead of trying to figure out dynamically a nice minimum size.  We are able to do this because I changed the square generator to increase the square size even if the current size is adequate.
This is much nicer in the case of an empty (or nearly empty) 
mask -- we'll just end up grabbing the entire image.

It is also much nicer for the images that are nearly completely
masked out, we're able to set the minimum size to a value that
we determine will produce a good fft Match - currently 512 but
I think this can be smaller.