A cleaning up of the water masking code.
Grow seed regions more slowly, so we can get more in the
end - they weren't powers of 2 anyway, and it would be
inconvenient to make them so.

Instead of using the unmasked-percent as the criterion for
which seed regions should be chosen, use the number of pixels
that are unmasked.  This allows us to pick larger regions, even
though they may have more pixels that are masked.  (All regions
still have >80% unmasked.)