We apparently had this problem of initializing the map projection block for ScanSAR imagery without having a proper earth radius before. Had fixed that in revision 2895.
Took the fix out again in some other revision but cannot reproduce why
I moved the code around at that time.

Lesson learnt from this: WE BADLY NEED REGRESSION TESTING. Establish a
baseline and run a whole bunch tests in an automatic fashion to figure
out whether any of the metadata changes have any bad effects. I will
have to make a bunch of changes in the future to get the entire ALOS
data suite to work, so we better get this thing in place sooner rather
than later.