Finally figured out how to eliminate the streaking on the left edge of the terrain corrected images.  Single leading 0.0 values do not cause problems, however two leading 0.0 values before the real data starts was causing problems with the "hole plugging" code.  For example:
0 1000 1003 1002 ... -> 0 1000 1003 1002.... (ok)
0    0 1000 1003 ... -> 0  500 1000 1003 ... (not ok)

The hole plugging code would put in a half-height value into the
"hole", which was taken as valid by subsequent code.  The leading 0
is ignored in both cases.

Fixed this by extending the hole-plugging code to plug at the
left edge of an image (copying the first valid data over).