1. Replaced strrchr()based replacement of '.' with '\0' (remove file extension) with ASF utility findExt() instead.
2. Commented out asfPrintWarning in asf_import.c where it warned
of defaulting to the 'generic geotiff importer' ...there is no
generic geotiff importer.  The code now prints a warning about
finding an unidentified type of geotiff and continues on.

3. Added error checking on the reading of the .aux file file
header to make sure the .aux file is valid.  Added similar
error checking to the reading of the data header to verify that
the version of the ArcGIS aux file is correct, else exit.