Add "detect_flag_options" to the asf library.
I've been meaning to take a bunch of the stuff I wrote for
command-line processing in libasf_geocode/parse_options and make it
available in the basic asf library; it just needs a little cleaning,
a little generalization, and some documentation & examples, and it
could be somewhat useful.

Since Rudi is itching for detect_flag_options I've added that one
now.