Time to check in the monster (before I completely forget what all I did).
Let's talk metadata first. I changed the band_number parameter to
band_count. Haven't used it, so should not be a problem. Added an
acqusition date parameter for the general block. Couple students in
class asked me how to find that in the metadata. It reached the
critical mass. I added a transform block that contains parameters for
coordinate transformations. That can be used to geocode optical data.

Added the ALOS style map projection record.

The rest of the changes is filling in all the possible metadata in the
respective blocks. For Palsar we have beam modes now. The optical
block for AVNIR and Prism is complete now. For optical images I
reactivated the location block. Makes a couple things easier.

The meta_get_latLon function can handle optical data now using the
parameters in the transform block.

Had to adjust the reading and writing functions to handle all the new
metadata.

Bottom line is that georeferenced and map projected ALOS data (all
flavors) can make it through the ingest, geocoding (for the
georeferenced images) and exporting to GeoTIFF.