The -band option enabled version of asf_export in all its glory.
Actually turned out to be a bit of work. Revived the
'extract_band_names' function to get the names for all available
bands. Removed the previous assumption for the format. Needed a
function to find a single band in a multi-band image, hence the
'find_single_band' function.

In order to be able to safely attach the band names to the output name
I wrote 'append_band_ext'. This function checks for known extensions
and strips those off to get the basename. After adding the band name
to the basename it adds the extension back again.