All the good stuff to export to PPM with the same function as TIFF, GeoTIFF and JPEG. Tossed the old functions out.
Looked into the look up table business, e.g. for color interferograms
and phase unwrapping masks. I guess we want to extend asf_export with
a -lut option that can read a look up table. If we stick the
predefined ones somewhere in the share directory, the users could add
theirs. Once the GUI is up to speed on that it could pick up all of
them, including the user defined ones.