Implemented look up tables handling.
Look up tables are text file (with a one line header) that contain 256
command delimited values for red, green and blue used for mapping
image values to a certain color. I have created two LUTs so far, one
for interferograms and one for phase unwrapping masks. They live in a
subdirectory 'look_up_tables' within libasf_raster and are copied into
the share directory at build time. I guess we could generate a LUT for
the terrain correction mask. Would sure be easier to analyze.

Had to reshuffle asf_export a little bit to accommodate the new
functionality. 'export-band.c' became a little lengthy so I split up
the writing lines functions and all the functions related to keys into
separate files.

Updates the call in libasf_convert.

The byte value conversion seems to work just fine. The floating point
version still has an issue with the downscaling. Could not spot it
right away. Though I check it in anyway.

Once we get the floating point version rolling, we can retire the DDR
structure for good. 'ardop' can produce the RGB output more
gracefully. We might want to think about the usefulness of PPM as an
output format. It is certainly no space saver and JPEG is certainly a
lot smaller. If you really want to have it without compression, go
with TIFF.