The revamped gui, based on Rudi's idea.
New "Completed Files" section that holds the successfully processed
files.  Got rid of old log window, which was klunky, the logs are
now accessed using the "View Log" button, formerly the jump-to-log
button.  Two toolbars instead of one, but the only new button
functionality is a "Queue for Reprocessing" which just means move
the file from the completed section back up to the not-yet-processed
list.

Replaced all glade_xml_get_widget() calls with a wrapper, from the
ait code, get_widget_checked(), that aborts on failure.  Much easier
to detect when the strings don't match between code & .glade file.