Moved airsargeo into unsupported, conveniently the last tool that depends on asf_geolib. The main reason for this move is that it is broken. It compiles just fine but a closer look reveals that it is completely relying on an old metadata structure that does not exist anymore.
If somebody wants it back (and that may be the case pretty soon with
ASF taking over the AirSAR archive), it needs some work. Not really
substantial work but I would look into a proper integration in
libasf_geocode. It uses some kind of grid for the mapping
function. Sounds quite like the approach of optical ALOS data.