There's a bug in glade with GtkComboBoxEntry -- it doesn't preserve the entries.  After setting up the radar selections (HH, VH, etc), and saving, the next time you'd start up glade, those entries would be wiped out for all GtkComboBoxEntry widgets, except the first one (which in our case is the "Red" -- the Green and Blue selections would be blanked, though it would leave the first item (HH)).
I'd been working around this by re-entering green & blue items every
time I edited the glade file.

The obvious workaround for this would be to set the items
programatically when convert starts, however this didn't work -- it
would core dump whenever I'd try to click on the any of the
GtkComboBoxEntry widgets.

But I've finally solved the problem.  According to one of the glade
developers:

> Heh,
>    you've definitly stumbled upon one or glade/libglade's wierder
> glitches ;-)
[...]
> So the confusing bottom line is that, libglade will setup you combo-box
> for regular gtk_combo_box_new_text() type interaction *if* there were any
> items set in the glade file, otherwise you just get a naked combo-box and
> you'll have to set it up yourself
[http://lists.ximian.com/pipermail/glade-users/2006-February/002864.html]

So, the solution is to leave the combo boxes *almost* blank -- put in
a "dummy" item in with glade, and programatically remove the dummy item,
and then add in the items that we actually want.