New library function, to_sr().  This converts an image to slant range.  Ground range images are given to gr2sr, so that works like before.  Projected images, as well as "transform"-style georeferenced images (i.e., ALOS), go through another new function: proj_to_sr(). (Slant range images are left unchanged.)
proj_to_sr() relies on meta_get_timeSlantDop() to do the heavy
lifting.  If we have the state vectors correctly hooked up (and I
think we do??), then I believe this will work on ALOS & Scansar.

Added a small command-line wrapper program to_sr that is used to test
the thing.

This should work on Scansar and ALOS images, though more testing is
needed.

When it does work, this will allow us to do terrain correction of
both of those kinds of images.