Latest fixes to the projected-to-slant conversion.
We now keep the azimuth time per pixel the same, if we have it.

Terrain correction of ALOS data now seems to work, provided you use
-no-resample.  (The resampling code doesn't properly update some of
the slant range pixel sizes, etc, in the metadata.  Fixing that
next, but I want to check this in to try it on Windows.)