On Linux I've been seeing some strange behavior.  waitpid() is used in the parent process to detect when the child process (which is running asf_convert()) completes.  However, the strange behavior is that this waitpid() doesn't successfully detect this completion.  The child has reached and run the exit() line, but the parent keeps waiting.
So I'm putting in a kludge to handle this situation.  The status file
that the GUI uses to update the Status column ("Importing..." etc) is
populated with "Done" when the child finishes.  If we find "Done",
then we'll just kill the child process (which is now a zombie) and
continue as if it had exited normally.