New scheme for interpolating over holes in the DEM.
Instead of doing it in reskew_dem, added new code to do it in a
separate step, libasf_sar/interp_dem_holes.

Added a small wrapper program "fill_holes" that runs the interpolator,
for testing or for users to fix their own dems prior to terrain
correcting, however the "-smooth-dem-holes" option to asf_terrcorr,
and the corresponding config file option, still work and do use this
new method.

The new method is really simple.  Iterate through the DEM, look for
points below the height cutoff (-900m by default), and sets the pixel
to a new value equal to a weighted average of the four nearest points
(up/down/left/right) that are "good" dem values.