Now require Gtk 2.6 for the GtkFileChooser.
Hurts to do this, since a lot of ASF machines have Gtk 2.4, and the
file chooser is pretty nice.

However, there's a bug in the file chooser in Gtk+ 2.4.x that causes
threading problems.
http://bugzilla.gnome.org/show_bug.cgi?id=161604

Apparently, this bug is fixed in 2.6.1:
http://mail.gnome.org/archives/gtk-devel-list/2005-January/msg00032.html

Our symptoms were that once we had called
gtk_file_chooser_dialog_new(), waitpid() could no longer successfully
detect if the child process had completed or not.

NOTE:
I haven't actually tried this on Gtk 2.6 to verify that this fix
really does fix this problem, I have a 2.6 box at home that I'll try
it on.