A last-ditch attempt to save the GtkFileChooser on Gtk+ 2.4.
You will surely recall the last check-in (to asf_convert_gui.h,
in svn rev. 5399), where I had discovered that there was a bug
in pre-2.6 GtkFileChooser where multi-threaded apps could behave
weirdly, and that we were consistently seeing this problem,
and so I bumped the version requirements for the chooser to 2.6.

A workaround for this bug is to find another way to find out if
the process has finished, and I've implemented that here -- using the
status file.  We now update the status file even when exiting with an
error, so that this kludge can terminate the child thread manually
if necessary.

The only holes in this plan:
1) Won't catch the case where the child thread core dumps.
2) Not entirely clear from the Gtk bug report what other symptoms of
this bug could be -- perhaps there are other problems that we haven't
yet noticed...

Anyway, let's try this out, see if we can salvage the Chooser.