Make proj_to_sr multiband aware.
The first band acts as the "reference" -- the mapping from projection space to slant range is determined by the contents in the first band, the other bands will use the same mapping.

Also, change the way the output pixels are stored.  Since this occurs line-by-line, there's no need to use the overhead of a float_image to store these, we can just use regular fopen/put_float_line with a single-line floating point array, and use much less memory & disk-io.