Redo how deskew_dem handles the DEM and the Mask.
Previously, the entire DEM was loaded in, so that column interpolation
could be performed.  This code was in the original deskew_dem,
presumably as a hole-handling method.  However, we have since added in
hole-filling as a pre-processing step (since SRTM DEMs have holes
larger than this column interpolator can handle).

So there is no longer any need to read in the entire DEM.  Changed to
read it in line-by-line.  (2-line-by-2-line, actually, we need current
& previous lines.)

There really was no reason to read in the mask all at once, but it is
now also done line-by-line.

Some other cleanup.