Major change in the proj library !!
Now all projection parameters must be passed in with degrees, not
radians.  The current code was sort of a mish-mash, some functions
would expect degrees, other radians.  Should now be completely
consistent.

However, it is a inconsistent with the actual lat/lon values you'd
like to convert (or get back) -- those should still be in radians.

This inconsistency is consistent with libproj itself...