Changes to get rid of the smearing problems near the edge of a DEM.
There were multiple factors that led to this.  Most of them boil
down to this: when interpolating, if ANY point used to interpolate
is a "no data" or "bad data" value, then the interpolation should
be skipped and you just return no data.  This was suggested by Orion
for asf_geocode.

First, remap wasn't obeying this rule -- now if any pixel value
matches the "background", the bilinear resampler now returns the
background value.

Second, reskew_dem was sometimes interpolating between "no data"
and "data" values, creating a long "ramp" in the slant range DEM.
This is basically the same problem, not ensuring all data used for
interpolating is valid.

Third, reskew_dem wasn't handling the case where first a "no data"
value maps to a slant range pixel, and a later "good data" pixel
on the same line maps to that same slant range pixel -- in this
case, the "no data" value would be left in the slant range DEM,
leaving "holes".

Fourth, deskew_dem was also introducing ramps by not checking for "no
dem data" on both ends of interpolating, when mapping the DEM from
slant to ground range.