Support for taking SLC data to amplitude/phase in asf_convert.
Added a [C2P] block, with just one parameter so far: multilook (boolean, if true then data is multilooked).  Calls the new c2p code, which generates a 2-band output image.

Just like the ardop block, it is a "secret" option, meaning that it will only appear in written config files if it is already set, so a user who isn't aware that the option can be added won't ever see it.