Major reorg/redo of libasf_geocode.  Moved code to a new function "asf_mosaic" which takes multiple input files, and geocodes each into the same output file.  To do this, had to split up the process into two phases: (1) determine extents, and (2) remap.  In between, we set up the output metadata.
Old asf_geocode_ext() is now a wrapper for asf_mosaic.

Made the command-line mosiac tool use asf_mosaic, which meant that mosic had to process all of the geocode command line options.  Moved the -help info on projection options to libasf_geocode so both could use the same text.

This still needs a good bit of testing.

When geocoding/mosiacking multi-band data, I decided to make the output have a number of bands equal to the largest input band count.  Any images with fewer bands won't contribute to the output in that band.

Another possibly controversial decision I made was, in the final resampling stage (where the output is produced), whether to loop over the inputs in the outer loop, and the bands in the inner - or vice-versa.  (Option A) Doing the inputs/outer-bands/inner means that we don't have to save or regenerate the splines produces to do the remapping, but we must keep all bands in memory (or in disk caches).  (Option B) Doing bands/outer-inputs/inner means we can write out each band as soon as completed, but must either save or rebuild the mapping splines.  I chose Option A, because it was easier, but we may want to revisit this, since I think B could be somewhat faster...