Implemented CEOS ingest for raw signal data from ESA and Beijing.
Still need to include the data sets into the test suite. Data seems to
implement fine but fail in ardop. Haven't found the reason for
that. Not sure whether that is related to import or not. Need to
include some more "regular" raw signal data to work this one out.