Implemented the ingest of AirSAR data as part of asf_import. The current implementation ingests data acquired in the ts_* and cm_* file naming scheme. This includes interferometric as well as polarimetric data.
For interferometric data the DEM, coherence and amplitude images are
stored in separate files.
For polarimetric data the compressed Stokes matrix is stored as a
nine-band image with total power as well as SHH, SHV, SVH and SVV (amp
and phase) as the bands.

No all data can be geocoded as some of the specific parameter files
are missing. Go figure.

In order to get the geocoding to work the peg point transformation
scheme needs to be included into meta_get_latLon (Kirk volunteered for that).