Implemented the multimatch2vector tool.
The functionality is all in the libasf_vector library with a little
wrapper that sorts out the parsing of the text file.

There are still issues with a text file. There are two entries for
some of the locations. That is all in Rick's court to tell us what to
do with this.