This is the look-up table ability that Rudi has been asking for.  In the viewer, you can choose one of the files from (sharedir)/look_up_tables (determined dynamically, like convert's canned projection list), and this will be applied to greyscale float & byte images.
However, it is limited by the fact that our lut code only works for data ranging from 0-255.  This works great for layover/shadow masks -- but for DEMs, for example, this isn't so great.

We don't want to use 2-sigma scaling before applying the lut (on layover/shadow masks, for example, that would result in incorrect color mappings), even for dems, you'd want the same elevation to always map to the same color.

We need to expand our look up table functionality to allow a greater input data range...