Implemented basic functionality to ingest GAMMA ISP data. Initial check in. Still needs some work on a number of polynomials. There must be another general parameter file that keeps track of basic information such as sensor, orbit, frame and the like.
Kept the real testing for the conversions between GAMMA and metadata
until the point when we got the MSP data in.

The writing of the GAMMA ISP parameter seem to be fine. Looks ok at a
first glance. The real test is taking some data set that went through
convert and see what the GAMMA software is able to do with it.