Change the projection-to-slant code (used in terrain correction of scansar & ALOS images) to always put the latest time on top, earliest time at the bottom.
Previously, I'd been using azimuth_time_per_pixel's (atpp) sign to determine this: if atpp>0, the image would be increasing in time with line number, if atpp<0, the image would be decreasing in time with line number.

Now, we always decrease time with line number.  If necessary, the sign of azimuth time per pixel is reversed.