AirSAR support in asf_convert.
New block "[AirSAR]" in the configuration file, which controls which
items are processed.  If some products are not present, then
further processing of those is turned off.

Pauli/Sinclair are allowed with the P/L band components of AirSAR,
these are turned off when doing the C band stuff.  Since the
polarimetric components of AirSAR are different than Palsar (has a
"Power" band), this ended up being a lot of special-case code.