Added .png and .pgm/.ppm support to asf_view.h
The .png support isn't able to handle all types of pngs, however it should handle everything we produce.  So, we may just be ready to call this done as well.  (We certainly won't be handling all types of pngs -- asf_view probably won't ever handle rgb-alpha...)

The pgm/ppm support is complete.