Sortable columns in the grid.  Orbit/frame numbers are in there, but the calculation is bogus at this point.
Checkboxes in the grid.  (But checkboxes don't do anything yet -- they
will turn on/off the corresponding acquisition on the map.)

A "save" button, saves selected columns.  But, probably, with the new
checkboxes, this should change to use the checkboxes.  And, actually,
"Save" doesn't actually save anything, just prints, but we have yet to
decide how results should be handled.

Support for looking at lat/lon pseudoprojected images -- this is
likely how the planner's map will be stored.