We have a bug in UTM2latLon() that I'm not really sure how to fix.
We were setting the False Northing value to 0 or 10^7, depending on
the value of "lat" as follows:
  false_northing = lat>0 ? 0 : 10^7

seem reasonable, eh?  Unfortunately, the function is calculating the
latitude, and "lat" is a pointer, which will always be >0.  So, the
code currently always sets this value to 0.  (Weird that this isn't
a compiler warning: perhaps -Wall doesn't turn this one on?)

So, I've replaced this line with just "false_northing = 0" so it
doesn't look so crazy, but this means that we don't get the right
lat/lon values in the southern hemisphere.

Fortunately, this code doesn't seem to be used much, so I 
don't *think* this is currently causing any major problems...