A start on the convert2vector gui.
For the first iteration, only one conversion will be supported,
which is converting from an ALOS CSV file (something you get from
downloading from a certain page on the JAXA site) to KML.  This
is something that Michelle has a specific need for by next week.

Eventually, we'll put all of the conversions supported by
convert2vector in here.

At this point, the GUI is sort of complete, and some of the event
handlers are in place, and the actual processing code has
been started.