The beam_modes.txt file now supports two formats- (1) [the previous one] where a min/max/increment are specified for the allowed look angles, and (2) [the new one] a list of allowable look angles.
In the case of (1), an entry box is shown in the GUI when that mode is
selected, and the user types in the look angle.  When a mode using
style (2) is selected, a drop-down list of the allowed beam modes is
shown.

Added Prism and Avnir selections.

Added support negative (left-looking) look angles.