Remaining code necessary to handle terrain correcting calibrated images by adding an amplitude band in the beginning, just for the coregistration.
What happens is: if the user elects to terrain correct non-amplitude
data (sigma, beta, gamma, power), then in addition to the usual bands
we also import an amplitude band as band #1.  This band is the
"reference" band during terrain correction (we already have set it up
so that terrain correction uses the first band as the reference during
coregistration), and afterwards it is removed.