Will not get to this much before the end of sprint, hence this check in.
Major rewrite of the convert2vector functionality. Has some
significant improvement.

The formats that are internally supported are:
metadata files
leader files
point files
polygon files
generic csv files
auig files
geotiff files
kml files
shape files

The main enhancement is the introduction of a header configuration
file 'header.lst' that lives in the share directory. This allows the
user to influence which parameters, e.g. for the metadata structure,
are actually included. A simple '#' at the beginning of a line
indicates that this parameter is to be handled as 'invisible'.
In a KML file these parameters are still included but commented out
(HTML style <!-- ).

The other major advantage that I just added is the possibility for the
user to define their own formats and pass them into convert2vector. So
far I have only hooked up a 'custom2shape' function. So for example,
if you define a [UTD] format in the 'header.lst' file with a couple
parameters in there (parameter name and format), you can pass now an
input format 'UTD' into convert2vector. The only other thing required
is an identifier in the data file that you pass in. The first line
should be

# Format: UTD

Plenty testing left but I thought you might want to check it
out. Feedback is always appreciated.