Added the ability to specify an rgb look up table for asf_export in the asf_mapready configuration file.
Added code to do the Cloude-Pottier classifications.  The classifications are done with configuration files, and map regions in entropy/anisotropy/alpha space to greyscale values.  Added some placeholder classification files, however we need to look up the real classifications.

Added look-up tables for the classified images, and hooked up asf_mapready & mapready to use the classifications and rgb look-up tables in order to make the gui options for cloude-pottier work.

Added a 2-D histogram in entropy-alpha space.  The polarimetry code dumps "ea_hist" img & meta files.  You will all be happy to hear that these histograms look quite like the ones in the papers!!  We get the sideways parabola looking thing, with a large clump of values on the bottom curve of the parabola.

Changed the asf_calpol command-line to handle a classification file on the command-line.

So, for now, I've turned back on the Cloude-Pottier stuff in the GUI.  Let's see if we can get this thing to work.