These are the changes needed to correct the time/slant offsets using corner & center point information found in the workreport file, for Palsar SLC data.
The code minimizes the geolocation error by comparing the results of
meta_get_latLon() with the corner/center point info in the workreport
file, by performing a 2d minimization: time/slant offsets -> total error.

The results seem pretty good so I'm checking this in, for 2.0.3 we
will do it this way.  However, we should continue to look in the CEOS
metadata to see if we can get better timing info from there.  (It is
the time offset that we're really trying to figure out, the
minimization seems to generate very small slant offsets.)