Added -save-metadata and -patches options.
The -save-metadata flag results in the copying of
all metadata files (.meta) created along the way
into the output directory.  Note:  When generating
a CEOS (non-Level 0) thumbnail, no metadata was
written out.  So in this case, meta_write() is 
used first, then the resultant .meta file is copied.

The -patches option is identical to the -p or -patches
option to ardop.  It only applies to Level 0 files.
It results in only <n> patches being processed by
ardop, i.e. for debug.  This option is NOT part of
the help output, i.e. it's a secret command line
option for internal usage.

=> Final testing in progress.  Initial testing was
   all good.