We want to add the amplitude band whenever we are doing terrain correction, and not using amplitude radiometry.  This will automatically pick up the polarimetry case, since there we hard-code to Sigma - so, turning on polarimetry and terrain correction will always add the amplitude band.
Import used to ignore the amp0 flag if the selected radiometry was already amplitude, this behavior is now changed -- the amplitude band is added anyway.  This is so that after Faraday Rotation correction, the 8-bands of SLC data can be calibrated, and then the 9th band (which is actaully the first, 0th, band in the file) will be passed through, ready for use by terrain correction.

So, the whether we turn on the amp0 flag doesn't depend on whether we're doing Faraday Rotation at all - just if we are terrain correcting non-amplitude data.