Modified the data handling of AirSAR data. When processing interferometric data, we ignore any polarimetric data (they are not in the same geometry).
The other addition is the checking of file existence before processing
(ingest, geocoding, export). If a user copies olny a part of the data set
for processing, the system should not just die. Switching off the not
existing data sets in the metadata might be a handy work around but
does not really cut it.