Start of the little side project.
This tool is supposed to synchronize flight information (image time
stamps, GPS and inertial measurements).

Implemented the first step, extracting the time information out of the
image data (JPEG format) using the EXIF library functionality. For the
moment the tool read the list of image files as input. It writes the basename
and acquisition date/time into a CSV file.