Fix for mantis 951.
Added code to collect the generated output filename(s) in asf_export.
Pass this up to asf_convert, and in asf_convert store these in the
"intermediates" file that is generated.

In the GUI, grab the first output filename (which is the one that was
used to generate the thumbnail) and use that in the completed files
list.

Should be able to extend this to generating one completed file entry
for each output band, but we'd have to sort out the thumbnails.
Save that for 2.2