Use the workreport to calculate the azimuth time per pixel for SLC data, instead of using the swath velocity calculation (which is actually a nadir velocity calculation, and is why we have scaling errors).
(This isn't needed on 1.5 data because there we use the transform
block for all geolocations.)

So, we now depend on the workreport for two things, the corner points,
and the scene timing information.

Fix workreport.c to be able to find the workreport even when given
the leader file name (instead of the basename).