Changes required for ingesting CEOS data flavors from three more groundstations (Tromso, Dera, Westfreugh).
Needed an early check in for Andreas' ship detection project. Still
working on the metadata related issues. Currently the imagery ingests
fine, so he has the images to look at with basic metadata (resolution)
available. However, no geocoding is possible.

Changed the unknown calibration scheme to a warning for the moment.