Update the MapReady documentation to 2.1.
Not added yet: AirSAR.

Not sure if we really want to add that, though - particularly for this beta release next week - since the support is still a bit rough... maybe we'll get it in shape for the November release, if so we can add it to the manual then.

So, I think the manual as it is now is probably good enough for the release next week.