Take out the meta_create() call in open_shape().
This should solve problems where open_shape() is used to open an
existing shapefile (instead of a new one).

The ".dummy" kludge needs to be applied because SHPOpen() will
always strip off an extension, so we need to ensure that one is
always there.  We can do this without involving meta_create,
just by using findExt().